requirements.txt
black==21.6b0
discord.py==1.4.1
nanoleafapi==1.1.5
PyQt5==5.15.4
python-dotenv==0.19.0
